service_olioli/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pillow==10.1.0
numpy==1.26.0
redis==5.3.1
pydantic==2.11.7
pydantic-settings==2.10.1
transformers==4.53.2
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1